Added a new hirerarchy for types, in a new sub package of types. continued changing MITRE categories
