+++ kube_hunter/core/events/types.py
@@ -7,8 +7,8 @@ from kube_hunter.core.types import (
     DiscoveryCategory,
     DenialOfService,
     RemoteCodeExec,
-    IdentityTheft,
-    PrivilegeEscalation,
+    LateralMovementCategory,
+    PrivilegeEscalationCategory,
     AccessRisk,
     UnauthenticatedAccess,
     KubernetesCluster,
@@ -105,8 +105,8 @@ class Vulnerability:
             DiscoveryCategory: "medium",
             DenialOfService: "medium",
             RemoteCodeExec: "high",
-            IdentityTheft: "high",
-            PrivilegeEscalation: "high",
+            LateralMovementCategory: "high",
+            PrivilegeEscalationCategory: "high",
             AccessRisk: "low",
             UnauthenticatedAccess: "low",
         }
