+++ kube_hunter/core/types/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa: E402
+from .hunters import *
+from .components import *
+from .vulnerabilities import *
