+++ kube_hunter/core/types/vulnerabilities.py
@@ -1,0 +1,54 @@
+"""
+MITRE ATT&CK Categories
+"""
+class InitialAccessCategory:
+    name = "Initial Access"
+
+class ExecutionCategory:
+    name = "Execution"
+
+class PersistenceCategory:
+    name = "Persistence"
+
+class PrivilegeEscalationCategory:
+    name = "Privilege Escalation"
+
+class DefenseEvasionCategory:
+    name = "Defense Evasion"
+
+class CredentialAccessCategory:
+    name = "Credential Access"
+
+class DiscoveryCategory:
+    name = "Discovery"
+
+class LateralMovementCategory:
+    name = "Lateral Movement"
+
+class CollectionCategory:
+    name = "Collection"
+
+class ImpactCategory:
+    name = "Impact"
+
+
+""" 
+General Categories 
+"""
+class UnauthenticatedAccess:
+    name = "Unauthenticated Access"
+
+class AccessRisk:
+    name = "Access Risk"
+
+class PrivilegeEscalation():
+    name = "Privilege Escalation"
+
+
+class DenialOfService:
+    name = "Denial of Service"
+
+
+class RemoteCodeExec:
+    name = "Remote Code Execution"
+
